+++ pkg/executor/task_executor.go
@@ -58,7 +58,13 @@ func (e *taskExecutor) Exec(ctx context.Context) error {
 	}
 	// exit when task run failed
 	if e.task.IsFailed() {
-		return errors.Errorf("task %q run failed", e.task.Spec.Name)
+		failedMsg := "\n"
+		for _, result := range e.task.Status.HostResults {
+			if result.StdErr != "" {
+				failedMsg += fmt.Sprintf("[%s]: %s\n", result.Host, result.StdErr)
+			}
+		}
+		return errors.Errorf("task [%s](%s) run failed: %s", e.task.Spec.Name, ctrlclient.ObjectKeyFromObject(e.task), failedMsg)
 	}
 
 	return nil
